requirements_linux.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
arviz=0.16.0=pyhd8ed1ab_0
binutils_impl_linux-64=2.40=hf600244_0
binutils_linux-64=2.40=hbdbef99_0
blas=2.117=openblas
blas-devel=3.9.0=17_linux64_openblas
brotli=1.0.9=h166bdaf_9
brotli-bin=1.0.9=h166bdaf_9
brotli-python=1.0.9=py311ha362b79_9
bzip2=1.0.8=h7f98852_4
c-ares=1.19.1=hd590300_0
ca-certificates=2023.5.7=hbcca054_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=5.3.1=pyhd8ed1ab_0
certifi=2023.5.7=pyhd8ed1ab_0
charset-normalizer=3.2.0=pyhd8ed1ab_0
cloudpickle=2.2.1=pyhd8ed1ab_0
cons=0.4.6=pyhd8ed1ab_0
contourpy=1.1.0=py311h9547e67_0
cycler=0.11.0=pyhd8ed1ab_0
etuples=0.3.9=pyhd8ed1ab_0
fastprogress=1.0.3=pyhd8ed1ab_0
filelock=3.12.2=pyhd8ed1ab_0
fonttools=4.41.0=py311h459d7ec_0
freetype=2.12.1=hca18f0e_1
gcc=12.3.0=h8d2909c_0
gcc_impl_linux-64=12.3.0=he2b93b0_0
gcc_linux-64=12.3.0=h76fc315_0
gxx=12.3.0=h8d2909c_0
gxx_impl_linux-64=12.3.0=he2b93b0_0
gxx_linux-64=12.3.0=h8a814eb_0
h5netcdf=1.2.0=pyhd8ed1ab_0
h5py=3.9.0=nompi_py311he78b9b8_101
hdf5=1.14.1=nompi_h4f84152_100
icu=72.1=hcb278e6_0
idna=3.4=pyhd8ed1ab_0
kernel-headers_linux-64=2.6.32=he073ed8_16
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.4=py311h4dd048b_1
krb5=1.21.1=h659d440_0
lcms2=2.15=haa2dc70_1
ld_impl_linux-64=2.40=h41732ed_0
lerc=4.0.0=h27087fc_0
libaec=1.0.6=hcb278e6_1
libblas=3.9.0=17_linux64_openblas
libbrotlicommon=1.0.9=h166bdaf_9
libbrotlidec=1.0.9=h166bdaf_9
libbrotlienc=1.0.9=h166bdaf_9
libcblas=3.9.0=17_linux64_openblas
libcurl=8.1.2=hca28451_1
libdeflate=1.18=h0b41bf4_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libexpat=2.5.0=hcb278e6_1
libffi=3.4.2=h7f98852_5
libgcc-devel_linux-64=12.3.0=h8bca6fd_0
libgcc-ng=13.1.0=he5830b7_0
libgfortran-ng=13.1.0=h69a702a_0
libgfortran5=13.1.0=h15d22d2_0
libgomp=13.1.0=he5830b7_0
libhwloc=2.9.1=nocuda_h7313eea_6
libiconv=1.17=h166bdaf_0
libjpeg-turbo=2.1.5.1=h0b41bf4_0
liblapack=3.9.0=17_linux64_openblas
liblapacke=3.9.0=17_linux64_openblas
libnghttp2=1.52.0=h61bc06f_0
libnsl=2.0.0=h7f98852_0
libopenblas=0.3.23=pthreads_h80387f5_0
libpng=1.6.39=h753d276_0
libsanitizer=12.3.0=h0f45ef3_0
libsqlite=3.42.0=h2797004_0
libssh2=1.11.0=h0841786_0
libstdcxx-devel_linux-64=12.3.0=h8bca6fd_0
libstdcxx-ng=13.1.0=hfd8a6a1_0
libtiff=4.5.1=h8b53f26_0
libuuid=2.38.1=h0b41bf4_0
libwebp-base=1.3.1=hd590300_0
libxcb=1.15=h0b41bf4_0
libxml2=2.11.4=h0d562d8_0
libzlib=1.2.13=hd590300_5
llvm-openmp=16.0.6=h4dfa4b3_0
logical-unification=0.4.6=pyhd8ed1ab_0
matplotlib-base=3.7.2=py311h54ef318_0
memory-profiler=0.61.0=pypi_0
minikanren=1.0.3=pyhd8ed1ab_0
mkl=2022.2.1=h84fe81f_16997
mkl-service=2.4.0=py311hb711fc7_0
multipledispatch=0.6.0=py_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.4=hcb278e6_0
numpy=1.25.1=py311h64a7726_0
openblas=0.3.23=pthreads_h855a84d_0
openjpeg=2.5.0=hfec8fc6_2
openssl=3.1.1=hd590300_1
packaging=23.1=pyhd8ed1ab_0
pandas=2.0.3=py311h320fe9a_1
pillow=10.0.0=py311h0b84326_0
pip=23.1.2=pyhd8ed1ab_0
platformdirs=3.8.1=pyhd8ed1ab_0
pooch=1.7.0=pyha770c72_3
psutil=5.9.8=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
pymc=5.6.1=hd8ed1ab_0
pymc-base=5.6.1=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
pytensor=2.12.3=py311hb755f60_3
pytensor-base=2.12.3=py311h320fe9a_3
python=3.11.4=hab00c5b_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd8ed1ab_0
python_abi=3.11=3_cp311
pytz=2023.3=pyhd8ed1ab_0
readline=8.2=h8228510_1
requests=2.31.0=pyhd8ed1ab_0
scipy=1.11.1=py311h64a7726_0
seaborn=0.12.2=pypi_0
setuptools=68.0.0=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sysroot_linux-64=2.12=he073ed8_16
tbb=2021.9.0=hf52228f_0
tk=8.6.12=h27826a3_0
toolz=0.12.0=pyhd8ed1ab_0
typing-extensions=4.7.1=hd8ed1ab_0
typing_extensions=4.7.1=pyha770c72_0
tzdata=2023c=h71feb2d_0
urllib3=2.0.3=pyhd8ed1ab_1
wheel=0.40.0=pyhd8ed1ab_0
xarray=2023.6.0=pyhd8ed1ab_0
xarray-einstats=0.6.0=pyhd8ed1ab_0
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.6=h166bdaf_0
zstd=1.5.2=hfc55251_7
